Allow unmanaged dependencies
Aether allows unmanaged dependencies to go through the CollectRequest
by setting a property on the Artifact.  This is similiar to the maven
"system" scope.  One use case is for tools.jar that ships with the jvm.
Unfortunatly, this does make it easy to add dependencies that will
not exist on other systems.

It is implemented by using :file metadata on the coordinates to
be consistent with other api methods.