Fixed #61: resolve-artifacts* type error when using a :files map
resolve-artifacts* accepts a map of known artifacts to file names but treats them as dependencies instead of artifacts. This causes a type error at runtime if a :files map is actually provided. This patch corrects the error (likely from copy-paste, see #61) and adds a test for resolving artifacts against a :files map.