Add support for `managedDependencies` - fixes #72
This commit adds support for passing a vector
of `:managed-dependencies` to `resolve-dependencies`
and `resolve-dependencies*`.  This allows pomegranate
to support maven's built-in capabilities for determining
version numbers from a `managedDepencies` configuration,
e.g. from a parent project.

This work is intended to help make it possible to provide
better support for parent projects in leiningen.