Make `merge-versions-from-managed-coords` fn public
Prior to this commit, the `merge-versions-from-managed-coords` fn
was marked as private, because I didn't think it would need to be
called from outside of pomegranate.

However, it turned out to be a requirement to be able to call this
from leiningen in order to implement support for managed dependencies
in leiningen, and the function seems like it could be generally
useful for other cases, so this commit makes it public.

In the current leiningen code we are simply calling the private
function, so getting this merged and release is not a blocker for
anything, but it'd permit a desirable minor cleanup in the
corresponding leiningen code.