Update wagon libs from 3.3.2 -> 3.3.4 (#128)
The motivation for this is to restore display of the reason phrase
returned from the server. Clojars relies on this to convey more
context to the user when a deploy fails. Displaying it was removed in
wagon-http 3.2.0, and restored in 3.3.4 due to user complaint.

This should help resolve https://github.com/clojars/clojars-web/issues/774
once this is released and lein is updated to use the new version of
pomegranate.

The changelog for 3.3.2 -> 3.3.4 is
https://issues.apache.org/jira/browse/WAGON-541?jql=project%20%3D%20WAGON%20AND%20fixVersion%20in%20(3.3.3%2C%203.3.4)
with WAGON-541 being the relevant issue covering the reason phrase
changes.