dev: add clj-kondo linting
Some clear clarity wins:
- Removed unused private vars
- Removed unused imported classes

Subjective:
- Unused destructured `:keys` and `:as` now are prefixed with underscore.
This allows these items to still act as docs, but conveys that they are not
used and more importantly not used on purpose.
- An if without else converted to when.
This very clearly conveys intent: no we did not forget the else condition.
- Converted :use clojure.test to :require :refer

Pure subjective liberties (not clj-kondo findings):
- I noticed that some nses were being required and not aliased.
To me this indicates the ns is being included for its side effects.
An alias, to me, conveys it is being used by code.

Closes #139